Next Permutation
question:
The next permutation of an array of integers is the next lexicographically greater permutation of its integer. More formally, if all the permutations of the array are sorted in one container according to their lexicographical order, then the next permutation of that array is the permutation that follows it in the sorted container. If such arrangement is not possible, the array must be rearranged as the lowest possible order (i.e., sorted in ascending order).

For example, the next permutation of arr = [1,2,3] is [1,3,2].
Similarly, the next permutation of arr = [2,3,1] is [3,1,2].
While the next permutation of arr = [3,2,1] is [1,2,3] because [3,2,1] does not have a lexicographical larger rearrangement.
Given an array of integers nums, find the next permutation of nums.

The replacement must be in place and use only constant extra memory.

 

Example 1:

Input: nums = [1,2,3]
Output: [1,3,2]
Example 2:

Input: nums = [3,2,1]
Output: [1,2,3]
Example 3:

Input: nums = [1,1,5]
Output: [1,5,1]
 

Constraints:

1 <= nums.length <= 100
0 <= nums[i] <= 100


solution:
explaination with example:
0 1 2 3 4 5 => index
1 2 3 3 2 1 => value in array
1.i have check the value last to first value.
1<2,2<3,3==3, 3>2
2.find the smallest value and then need to find greater than value for that finding value. so last to that smallest value.
3.we will reverse the all value if that no smallest value.
4.if find the smallest value, found smallest value is 2 and index is 1.
5. swap that index value. and then reverse the last to that smallest value of this array.